Polished the website
Tried to tie off all the loose ends.  The website and the GitHub project now match in terms of hymns being worked on and their state; and entries for the website are now linked to the Github issues.  All review instructions are in place and everything seems to render.  All the research is in, with supporting documents copied over.